Finished messenger
Sockets are working correctly and the server and clients are able to send and receive messages between each other. The double ratchet objects between each server-client connection is staying in sync, although further testing is needed to determine the strength of this sync. The user must enter the password set by the server at the command line in order to connect. When connected, a user many send a message that the server will broadcast to every other connected client. Sending "EXIT" will disconnect the client from the server. Any encryption method besides AES (AEAD) is not possible as of this moment since the python-doubleratchet and cryptography packages only seem to support AES (RSA and Twofish are not supported).